backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.1.0
cohere==2.3.1
python-dotenv